Added annoying spacing-line
Added the fucking annoying line-between-header-main design choice. Could (and tested it succsessfully) with vanilla HTML + vanilla CSS + typescript. The lack of typescript made this a 30% ChatGPT part and I hate it. I also hate the solutions I made and that I had to do magic in the line-.svg file as well. I hate this!